project-template/pyproject.toml
[tool.poetry]
name = "django-starter-template"
description = ""
version = "0.1.0"
authors = []

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
psycopg2-binary = "^2.9.3"
# Extensions just for werkzeug really, maybe unnecessary?
django-extensions = "^3.2.0"
whitenoise = "^6.2.0"

# Hosted deps
circus = "^0.17.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-django = "^4.5.2"
pytest-cov = "^3.0.0"
factory-boy = "^3.2.1"
Werkzeug = "^2.1.2"
flake8 = "^5.0.4"
pylint = "^2.14.5"
pylint-django = "^2.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
